tdf#105093 sd PDF export: handle embedded videos
In practie embedded files always have a temp file URL, so from the PDF
export's point of view they are still URLs, just at the end the contents
of the URL is embedded to the PDF, not just the URL itself.

So add a SetScreenStream() that's similar to SetScreenURL(), but it's
for embedded, not linked videos.

Change-Id: Ifcc60357ef0f5fed0bdec02e0c84cb16ee147781
Reviewed-on: https://gerrit.libreoffice.org/32727
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>