PivotTable: Avoid adding useless 'Data' column field to table layout
Layout dialog adds this column field by default, but it has
no actual meaning or functionality. It must be added because
of this broken logic here (adding a field only when its hidden?).
This 'Data' field must be related to OOXML format see
OOX_PT_DATALAYOUTFIELD, but might be deprecated actually.

Change-Id: I90f00ef6290b343f84e40e5c94fd0e29a328329b