tdf#105077 sw: fix wrong assert corner-case in SwRegHistory::InsertItems
There is a tricky corner-case when an item set is inserted that
contains items that are all identical to what's in the node's item set.

In this case a SwHistorySetText will be created for an overwritten hint,
but no new AUTOFMT hint will be inserted so there is no
SwHistoryResetText.

So it looks like AUTOFMT can't be checked at all here.

(regression from 5b3b8117c38a64ca48e758011d3098d435416ec4)

Change-Id: I83a241588cef6a9ec074d4f0e39322987394ca86