Always exit with error if codesign fails
That sanity check was added in
9607f9edce46b4614628395a8fffbd7ed679ce4a but we lost it at some stage.

The codesigning script has a very confusing history, in part caused by
its use for two purposes, when building for the Mac App Store and when
building for a TDF style distribution on a dmg disk image. Those who
work with the former and those who work with the latter haven't
necessarily checked that it doesn't break for the other case. Sorry.

Change-Id: I79011302f60b1f6551328c8b80e00f5d3698504c