tdf#105015: Don’t ignore all printer fonts
For whatever reason we query printer output device for fonts and for
whatever reason GDI will only report fonts from the printer which
often in formats we want to ignore. This results in aborting the
application when we reject all fonts on the printer!

Just special case printer output devices and hope for the best.

I think this commit is partially blamed as it is evidenced now that
there was a good reason to only abort for window output devices, but
I’m leaving this as it is for now:

commit 18e8302abbf1a38fef55abf661ad3685a22efe11
Author: Chris Sherlock <chris.sherlock79@gmail.com>
Date:   Tue Mar 25 21:54:20 2014 +1100

    fdo#74702 OutputDevice::ImplInitFontList() fails if no fonts on device

    It makes no sense that ImplInitFontList() only fails for Window
    instances. I have carefully checked all the functions that use
    this function, and there are no good cases when no fonts won't cause
    problems. In fact, we have a number of functions that specifically
    rely on the fact that ImplInitFontList will populate
    OutputDevice::mpFontCollection with at least one font.

    Therefore, I'm making this abort if it can't populate the collection,
    regardless of whether it is a Window, Printer or VirtualDevice.

    I have also refactored GetDefaultDevice - I now check the default
    pOutDev parameter to see if it is NULL (the default), in which case
    it is referring to the default window, so I call on
    Application::GetDefaultDevice() instead of going straight to the
    pimpl data structure used by the Application class.

Reviewed-on: https://gerrit.libreoffice.org/33014
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>
(cherry picked from commit 819e52e6b83fe43de28de2a2bd0b5d026311202d)

Change-Id: I3f34ba0cea9f88966ace3a5ef7fac868972ea5ac
Reviewed-on: https://gerrit.libreoffice.org/33087
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Khaled Hosny <khaledhosny@eglug.org>