tdf#105188 sd: fix shutdown crash after accessing master pages
The problem here is that the destructor of SdModule does a lot of
things, including destroying an SdXImpressDocument that is referenced
from some SdGlobalResourceContainer.

This calls SD_MOD() to get the SdModule to get some resource, but at
that point SfxApplication::GetModule() returns null, because the
sequence was changed from first deleting the SfxModules, then clearing
the pointer in ~SfxModule to null, to unique_ptr::reset(), which, at
least in libstdc++, is implemented via std::swap, so it clears the
pointer before deleting the SfxModule.

It appears rather brittle to rely on such a subtle detail, so refactor
things so that SdGlobalResourceContainer is no longer owned by SdModule
but has its own pet XTerminationListener, which means it will be
destroyed earlier, while the SdModule is still fully alive.

(regression from 311077e2df872dd00c6d243063aaff810f3f9e3d)

Change-Id: I7f03f3adf431be8728ef3d65a078b536cb96f959