tdf#105259 calculate sizes of all area tab pages on construction
When we open dialog that contains the area tab, we need to
calculate the size of area tab pages (that are triggered by
clicking on a button) and set the page area to the appropriate
combined size. Otherwise we don't account for the needed space
correctly and some page will be squished or have overlapping
controls.

This change creates all pages at dialog construction, gets
all the optimal sizes and sets the size of the page container box
to the combined (minimal) size.

Change-Id: Ie04a121810b96973f6e4502a52af675b2baacf25
Reviewed-on: https://gerrit.libreoffice.org/33040
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Tomaž Vajngerl <quikee@gmail.com>