Enable Chart XShape tests on Windows by default
They pass both locally and on CI builds on Windows, so I think they
are reliable enough to be enabled by default. On Linux CI builds one
tests is failing but not locally, so not enabling them on Linux for
now. Mac is a whole different story with most of the tests failing.

Change-Id: I1f2cf6f318ddce3c68d7353c49fc510f895bbb6a
Reviewed-on: https://gerrit.libreoffice.org/33173
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Khaled Hosny <khaledhosny@eglug.org>