tdf#99074 sw: remove duplicate property "IsBrowseMode" again
Commit b0959d3438cea744296e1816a087db0a77733736 added the property
"IsBrowseMode" to SwXDocumentSettings, but it is already available in
the API as SwXViewSettings property "ShowOnlineLayout".

The problem is that both of these properties get exported in ODF into
settings.xml, so it contains "IsBrowseMode" twice.

Unfortunately the SwXViewSettings are not available in writerfilter,
because the XModel::getCurrentController() is null, the view is created
after the import.

But there is already a way to store ViewData in the SfxBaseModel, which
is then used by SfxBaseController::ConnectSfxFrame_Impl() when creating
the view.  This applies the property at just the right time.

Change-Id: I842845d09a7b3fe81e27a1ed8ac8a8594da7f4e8