tdf#104334 DOC import: don't map STYLREF with switches to SwChapterField
Commit 5ff482214761648e7885c7796b6060c6a12b77ca (tdf#99100 DOC import:
handle subset of STYLEREF natively, 2016-04-06) started to import the
STYLEREF field code as SwChapterField in case it had \[a-z] switches
after the field argument.

The bugdoc has a \s switch, which is not documented in ISO/IEC
29500-1:2012 17.16.5.59 "STYLEREF", nor is really clear what it would
do.

The original use-case was for the " STYLEREF 1 \* MERGEFORMAT " syntax,
so just keep using fieldmarks for the " STYLEREF <field argument> \[a-z]
" case.

Change-Id: Ibd094e72be4e179c779a79a75769b599d4d9bba8
Reviewed-on: https://gerrit.libreoffice.org/33489
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>