Resolves: tdf#104598 make bright theme shadow the same as it always was
The shadow color was originally set via StyleSettings::Set3DColors taking the
background color as the input, and doing DecreaseLuminance(64) on it to make
the shadow color

So, continue to follow that mode if the "dark shadow" is darker than the
foreground, and IncreaseLuminance the equal amount the other direction for dark
themes.

This way under a dark theme the problem of impossible to see rulers of
tdf#90214 is still fixed, while the bright theme colors are the same as they
always were, which avoid the wearing-on-the-nerves "regression" problem.

Change-Id: I33133ff180ecadce3a14ee45bd33db73c56c0225