Gbuild-to-ide, Froze non-active supported generators.
Copied gbuild-to-ide.py from 079b98547c8d1782bb0a189b73cb5ffb1b2556e7
to gbuild-to-ideNS.py

Changed Makefile.in to use gbuild-to-ideNS.py for non-actively
supported platforms.

That way these platform are stable, and not affected of new features.

For now all platforms are still present in gbuild-to-ide and will
have a "does not break" test for all new features.

Change-Id: I8bb069d93c954a80de540a32b8a01889035c840d
Reviewed-on: https://gerrit.libreoffice.org/33710
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: jan iversen <jani@documentfoundation.org>