tdf#103976 DOCX import: disable incomplete w:before/afterLines style handling
Regression from commit abf0b6e6f01b149fb7b5298448da36bd8ff18335 (DOCX
import: parse <w:spacing>'s w:before/afterLines attribute, 2014-10-17),
the problem is that OOXML has 3 different attributes for the paragraph
bottom margin (and other 3 for the top one), while in Writer we just
have a top margin.

Now the import filter tries to work out which one of these should have
priority and ignore the rest, but this is way more complicated when
style inheritance has to be taken into account as well.

To avoid the regression just restrict w:before/afterLines handling for
the case when it's used as direct formatting, that's why this was
introduced after all.

Change-Id: Ie8642c7a9771596def6b8899e098b26c4f8be0b4
Reviewed-on: https://gerrit.libreoffice.org/33738
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>