Restart the system timer if it returns to early
At least on Windows our GetSystemTicks() implementation - using
QueryPerformanceCounter - occasionally states, the timer returned
too early, which stops processing further LO events.

As a workaround we restart the timer, as it's now the only source
of running the LO main loop.

Probably we should use osl_getSystemTime, for measuring, which uses
GetSystemTimePreciseAsFileTime, if available. Nothing states
anything is monotonic, so in this regard it may be flawed too.

Change-Id: I94eacd8f5bacf277a68575cc8db84653cbc49d12