tdf#98665 optimize table format style access
Rework GetTableFrameFormat and GetTableFrameFormatCount to a
simpler implementation (searching forward and using c++11).

Using GetTableFrameFormatCount to get the size and then in a loop
call GetTableFrameFormat for every index, can get really slow as
in each call we need to filter the whole collection. Through UNO
we can't avoid this (without much more work), but for internal
calls like SwXTextTableStyle::isInUse, we access the underlaying
collection and iterate + filter ourselves. In the same way we can
slightly optimize SwXTextTables::getByIndex UNO method (with
removing the need to call GetTableFrameFormatCount).

Change-Id: Ib8462c32311ccc162ec290fe4eec70820855a378
Reviewed-on: https://gerrit.libreoffice.org/34008
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Tomaž Vajngerl <quikee@gmail.com>