tdf#106059 PDF export: write form XObject proxy for PDF images
PDF allows referencing a page of an existing (embedded) PDF file
together with a fallback bitmap, but this is only allowed for form
XObjects, not for image XObjects.

So case the image is a PDF one, wrap the image XObject in a form
XObject, that will allow later referring to the embedded PDF image.

Change-Id: I30839dd6696f1fddd8ca2caa0d33123e90fcedbc
Reviewed-on: https://gerrit.libreoffice.org/34376
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>